+++ scripts/consoleOperations/options/consoleOptionsLoader.js
@@ -19,6 +19,11 @@ export class ConsoleOptionsLoader {
       });
    }
 
+   static async loadClosedExhibits() {
+      const result = await ConsoleOperationsClient.getClosedExhibitOptions();
+      return result.exhibits;
+   }
+
    static async loadOffDisplayExhibits() {
       const result = await ConsoleOperationsClient.getOffDisplayExhibitOptions();
       return result.exhibits;